scripts/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.29.0
algoliasearch==3.0.0
